Swap from HDRP to standard pipleine - Material updates- puzzle update
Switched to URP because of most of the common shaders use URP, it should also help maintain a better framerate.
Downloaded additional shaders and uninstalled/reinstalled other shaders
Added new floor, an activation chain which consists of a pipe, a watertower, a drain button and a final interactable.
New materials for various objects.
A button in editor to automatically fill the player's moisture
Water level inside the player instead of as UI
NPCs have head tracking to look at the player

TODO:
NPC dialogue
A few practice puzzles with rewards
Ways to unlock additional nozzles
More enemies and better Ai for those enemies